Move EmojiListPage to emoji\views.py
